Replace ulong with uint32_t to have the same size across platforms
On 64-bit platforms ulong has different sizes:
32 bits on windows (LLP64) and 64 bits on most Unix platforms (LP64).

See also: LLP64 and LP64 data models.

+++ src/libdev/afx/resource.cpp
@@ -58,7 +58,7 @@ std::string AfxResourceLib::getString(const uint id) const
         return "";
 }
 
-ulong AfxResourceLib::bitmapHandle(const std::string& name) const
+uint32_t AfxResourceLib::bitmapHandle(const std::string& name) const
 {
 	return 0;
 }
